requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
opensearch-py>=2.0.0
tqdm>=4.66.0
urllib3>=2.0.0
tenacity>=8.2.0
numpy>=1.24.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
scipy>=1.10.0
hdbscan>=0.8.33
flask>=2.3.0
flask-cors>=4.0.0
pika>=1.3.0
pymysql>=1.1.0
redis>=5.0.0
